requirements.txt
pyyaml
python-dotenv
requests


# gemini-api需要的依赖
google-generativeai
jsonpath-ng